feat: add Makefile to project with pam, build (local and global) targets
This commit introduces a Makefile to the project, that can setup pam and
build local and global targets. This is necessary as without this, users
cannot understand or solve the major problem of the screenlock not
unlocking even if they give the right password. The Makefile is simple
to run and is highly flexible for future target addition like debug etc.